requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-filter==2.4.0
django-haystack==3.0
django-rest-swagger==2.2.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
drf-haystack==1.8.9
drf-yasg==1.17.1
elasticsearch==1.8.0
elasticsearch-dsl==7.3.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
msgpack-python==0.5.6
openapi-codec==1.3.2
packaging==20.4
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
pyelasticsearch==1.4.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
uWSGI==2.0.19.1
vine==1.3.0
wcwidth==0.2.5
zipp==3.3.0
